pre_requirements.txt
setuptools==57.5.0
anyjson
django-celery
